infra: update repo's workflows and basic template
Main things:
- Update used third-party actions to their latest
- Use ubuntu-latest as a runs-on field
- Cleanup ACT stuff, it wasn't used in a while.

Why ubuntu-latest instead of a some fixed version?

The idea is to avoid deprication events related to the Ubuntu versions.
The basic workflow is really basic, it doesn't do much and thus
it shouldn't need to require much maintenance over the years.

By tracking the latest ubuntu version we shed some of that maintenance
weight away.